Remove useless data.length check which always succeeds
Before the substring was even copied into “data”, we’ve already
verified that “queued.length” was at least dataLength+4 (since we start
at pos 4). Therefore, the message length has already been validated and
the substring copied into “data” will always be exactly as long as
dataLength, 100% of the time. This if statement is therefore useless
and should be removed.