Fix leftover ping, check for binary messages
When a socket is successfully opened, a ping is set up so MFC doesn't
close an inactive connection. When the socket unexpectedly closes,
however, or runs into an error, the interval was still trying to ping
the server which would result in an unhandled exception. MFCSocket will
now clear the interval on a socket error or close.

Also, originally MFCSocket was designed to work in a browser using a
browser websocket implementation. In the conversion to an NPM package,
the event listeners were not moved over to the EventEmitter interface
provided by WS. MFCSocket now uses the EventEmitter interface.

Finally, MFCSocket will now check to see if a binary message was
received and ignore it when it does.