Add configurability, modify to allow latest use of WS.
Add the ability to configure many of the constants used. To override a
default, add a mfcsocket.json file to the root of the project and set
any of the json properties shown at the top of MFCSocket.js.

Also limited the protocol to RFC6455 (configurable) and set an origin
when connecting to the websocket which allows us to use the latest
version of ws.